Backend/Base_backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pymongo
pydantic
python-multipart
PyMuPDF
easyocr
paddleocr
paddlepaddle
opencv-python
langchain
langchain-community
langchain-core
langchain-huggingface
faiss-cpu
sentence-transformers
groq
pytesseract
pillow
gTTS
wikipedia
